requirements.txt
flask
flask-socketio
pymongo
eventlet
bcrypt
requests
pyotp
PyJWT==2.8.0
Pillow